Cleaned up file organization, moved source files into src
