+++ index.js
@@ -1,7 +1,7 @@
 import express from "express";
-import directionsResponse from "./directions.js";
-import carResponse from "./car-search.js";
-import updateRoute from "./update-route.js";
+import directionsResponse from "./src/directions.js";
+import carResponse from "./src/car-search.js";
+import updateRoute from "./src/update-route.js";
 import "dotenv/config.js";
 
 // require('dotenv').config();
